feat:新增Mongo工具库

+++ Level0/Materal/Src/Materal.Utils/Model/PageModel.cs
@@ -8,15 +8,15 @@
         /// <summary>
         /// 总页数
         /// </summary>
-        public int PageCount => (int)Math.Ceiling(DataCount / (decimal)PageSize);
+        public long PageCount => (int)Math.Ceiling(DataCount / (decimal)PageSize);
         /// <summary>
         /// 开始序号
         /// </summary>
-        public int StartIndex => (PageIndex - PageStartNumber) * PageSize + 1;
+        public long StartIndex => (PageIndex - PageStartNumber) * PageSize + 1;
         /// <summary>
         /// 数据总数
         /// </summary>
-        public int DataCount { get; set; }
+        public long DataCount { get; set; }
         /// <summary>
         /// 构造方法
         /// </summary>
@@ -26,7 +26,7 @@
         /// </summary>
         /// <param name="pagingIndex">页面位序</param>
         /// <param name="pagingSize">显示数量</param>
-        public PageModel(int pagingIndex, int pagingSize) : this(pagingIndex, pagingSize, 0)
+        public PageModel(long pagingIndex, long pagingSize) : this(pagingIndex, pagingSize, 0)
         {
         }
         /// <summary>
@@ -34,7 +34,7 @@
         /// </summary>
         /// <param name="pageRequestModel"></param>
         /// <param name="dataCount">数据总数</param>
-        public PageModel(PageRequestModel pageRequestModel, int dataCount) : this(pageRequestModel.PageIndex, pageRequestModel.PageSize, dataCount)
+        public PageModel(PageRequestModel pageRequestModel, long dataCount) : this(pageRequestModel.PageIndex, pageRequestModel.PageSize, dataCount)
         {
         }
         /// <summary>
@@ -43,7 +43,7 @@
         /// <param name="pagingIndex">页面位序</param>
         /// <param name="pagingSize">显示数量</param>
         /// <param name="dataCount">数据总数</param>
-        public PageModel(int pagingIndex, int pagingSize, int dataCount) : base(pagingIndex, pagingSize)
+        public PageModel(long pagingIndex, long pagingSize, long dataCount) : base(pagingIndex, pagingSize)
         {
             DataCount = dataCount;
         }
